Review fixes 2
