+++ relayer/src/main.rs
@@ -13,6 +13,7 @@ use prover::proving::GenesisConfig;
 use relay_merkle_roots::MerkleRootRelayer;
 use utils_prometheus::MetricsBuilder;
 
+mod common;
 mod ethereum_beacon_client;
 mod ethereum_checkpoints;
 mod hex_utils;
